various bug fixes

+++ engine.py
@@ -23,7 +23,7 @@ class Engine(L.LightningModule):
 
     def training_step(self, batch, batch_idx):
         im_ids, samples, targets = batch
-        samples_encoded = self.clip_model.encode_image(samples)
+        samples_encoded = self.clip_model.encode_image(samples).to(torch.float32)
         concept_logits, class_logits = self.model(samples_encoded)
         
         if self.model.retrieved_concepts is None:
